client/requirements.txt
websockets==10.4
aiofiles==23.1.0
RPi.GPIO==0.7.1
pygame==2.1.3
